requirements.txt
altair==4.2.2
attrs==23.1.0
beautifulsoup4==4.11.1
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.2
fonttools==4.39.2
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.2
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==1.4.4
Pillow==9.3.0
plotly==5.22.0
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pydeck==0.8.1b0
pydyf==0.5.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.1
pyphen==0.14.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==13.3.5
selenium==3.141.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
streamlit==1.33.0
tenacity==8.2.2
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
Unidecode==1.3.6
urllib3==1.26.18
validators==0.20.0
watchdog==3.0.0
weasyprint==58.1
webdriver-manager==4.0.0
webencodings==0.5.1
wsproto==1.2.0
zipp==3.15.0
zopfli==0.2.2
